refactor: 달력 뷰에 대한 리팩토링 수행
- 컴포넌트 내부에 선언되어 있던 makeDayRow 함수를 utils로 이동
- makeDayRow 함수는 상수 FRONT_HALF, REAR_HALF 에 대한 참조를 가져 해당
  상수 역시 같이 이동시켜줌
- 오늘 날짜가 row의 중간에 위치하도록 수치 조정
